server/requirements.txt
requests
Django==3.1.3
Pillow
gunicorn==20.1.0
ibm-cloud-sdk-core==3.10.0
ibm-watson==5.2.2
ibmcloudant==0.0.34
django-environ